scripts/nlweb-demo/requirements.txt
azure-identity
openai
python-dotenv
azure-cli
azure-ai-projects>=2.0.0b1
